backend/requirements-production.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# LiveKit
livekit==0.1.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3
pendulum==2.1.2

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.4
flower==2.0.1

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# Encryption
pycryptodome==3.19.0

# Email
aiosmtplib==3.0.1

# Prometheus Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0 